+++ test/__mocks__/fs.js
@@ -18,7 +18,7 @@ function createReadStream(file) {
       await h.sleep(10)
       return x
     })
-    .resolve()
+    .mapAsync((value) => value)
     .flatten()
     .toNodeReadable()
 }