+++ test/generator.test.js
@@ -29,7 +29,7 @@ test('generator backpressure', async () => {
       expect(sourceStream.paused).toBe(true)
       return x
     })
-    .resolve()
+    .mapAsync((value) => value)
     .toArray()
   expect(res).toEqual([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])
 })